requirements.txt
alabaster==0.7.12
amqp==2.6.0
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==0.15.7
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
billiard==3.6.3.0
binaryornot==0.4.4
black==19.10b0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
coverage==5.1
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
distlib==0.3.1
Django==3.0.7
django-allauth==0.42.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==3.0.1
django-model-utils==4.0.0
django-redis==4.12.1
django-stubs==1.5.0
djangorestframework==3.11.0
docutils==0.16
factory-boy==2.12.0
Faker==4.1.1
filelock==3.0.12
flake8==3.8.3
flake8-isort==3.0.0
future==0.18.2
hiredis==1.0.1
identify==1.4.21
idna==2.10
imagesize==1.2.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
jinja2-time==0.2.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
mypy==0.782
mypy-extensions==0.4.3
nodeenv==1.4.0
oauthlib==3.1.0
packaging==20.4
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.5.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyGithub==1.51
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytest-sugar==0.9.3
python-dateutil==2.8.1
python-lokalise-api==0.1.0
python-slugify==4.0.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.0
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
testfixtures==6.14.1
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.25
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
